feat(nextjs-template): optimize mermaid load
