+++ packages/nextjs-template/types/window.d.ts
@@ -1,8 +1,11 @@
+import { Mermaid } from "mermaid";
+
 // make it a module
 export {};
 
 declare global {
   interface Window {
     currentTheme: "light" | "dark";
+    mermaid?: Mermaid;
   }
 }
